requirements.txt
grpcio
grpcio-tools
pika==1.2.0
requests